+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
+### Added
+- NetworkCircuits - Implement new endpoint `/networkcircuit/*`.
 
 ## [8.1.0] - 2024-01-15
 ### Added